+++ source/playbooks/CIS/lambda/cis22.py
@@ -39,6 +39,8 @@ APPLOGGER = LogHandler(os.path.basename(__file__[:-3])) # application logger for
 # Get AWS region from Lambda environment. If not present then we're not
 # running under lambda, so defaulting to us-east-1
 AWS_REGION = os.getenv('AWS_DEFAULT_REGION', 'us-east-1')
+# Append region name to LAMBDA_ROLE
+LAMBDA_ROLE += '_' + AWS_REGION
 BOTO_CONFIG = Config(
     retries={
         'max_attempts': 10
@@ -57,6 +59,7 @@ def lambda_handler(event, context):
     try:
         for finding_rec in event['detail']['findings']:
             finding = Finding(finding_rec)
+            LOGGER.info('FINDING_ID: ' + str(finding.details.get('Id')))
             remediate(finding, metrics.get_metrics_from_finding(finding_rec))
     except Exception as e:
         LOGGER.error(e)