+++ source/playbooks/CIS/lambda/cis24.py
@@ -43,6 +43,8 @@ APPLOGGER = LogHandler(PLAYBOOK) # application LOGGER for CW Logs
 # Get AWS region from Lambda environment. If not present then we're not
 # running under lambda, so defaulting to us-east-1
 AWS_REGION = os.getenv('AWS_DEFAULT_REGION', 'us-east-1')
+# Append region name to LAMBDA_ROLE
+LAMBDA_ROLE += '_' + AWS_REGION
 BOTO_CONFIG = Config(
     retries={
         'max_attempts': 10
@@ -61,6 +63,7 @@ def lambda_handler(event, context):
     try:
         for finding_rec in event['detail']['findings']:
             finding = Finding(finding_rec)
+            LOGGER.info('FINDING_ID: ' + str(finding.details.get('Id')))
             remediate(finding, metrics.get_metrics_from_finding(finding_rec))
     except Exception as e:
         LOGGER.error(e)
@@ -137,7 +140,7 @@ def remediate(finding, metrics_data):
     # Set name for Cloudwatch logs group
     cloudwatchLogGroup = 'CloudTrail/CIS2-4-' + non_compliant_trail
     # CloudTrail to CloudWatch logging IAM Role on for each account
-    cloudtrailLoggingArn = 'arn:aws:iam::' + finding.account_id + ':role/SO0111_CIS24_remediationRole'
+    cloudtrailLoggingArn = 'arn:aws:iam::' + finding.account_id + ':role/SO0111_CIS24_remediationRole_' + AWS_REGION
 
     # Connect to APIs
     try: