+++ source/playbooks/CIS/lambda/tests/test_cis22.py
@@ -28,18 +28,21 @@ from lib.applogger import LogHandler
 from lib.awsapi_helpers import BotoSession, AWSClient
 from lib.applogger import LogHandler
 import lib.sechub_findings
+import tests.file_utilities as utils
 
 log_level = 'INFO'
 logger = Logger(loglevel=log_level)
 test_data = 'tests/test_data/'
 
+my_session = boto3.session.Session()
+my_region = my_session.region_name
+
 def test_event_good(mocker):
     #--------------------------
     # Test data
     #
-    test_event = open(test_data + 'cis22.json')
-    event = json.loads(test_event.read())
-    test_event.close()
+    event = utils.load_test_data(test_data + 'cis22.json', my_region)
+
     sns_message = {
         'Note': '"Enable CloudTrail log file validation" remediation was successful',
         'State': 'RESOLVED',
@@ -86,4 +89,4 @@ def test_event_good(mocker):
     resolve.assert_called_once_with(
         'RESOLVED: "Enable CloudTrail log file validation" remediation was successful'
     )
-    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, 'us-east-1')
+    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, my_region)