+++ source/playbooks/CIS/lambda/tests/test_cis29.py
@@ -27,18 +27,20 @@ from lib.logger import Logger
 from lib.applogger import LogHandler
 from lib.awsapi_helpers import BotoSession, AWSClient
 import lib.sechub_findings
+import tests.file_utilities as utils
 
 log_level = 'info'
 logger = Logger(loglevel=log_level)
 test_data = 'tests/test_data/'
 
+my_session = boto3.session.Session()
+my_region = my_session.region_name
+
 def test_not_configured(mocker):
     #--------------------------
     # Test data
     #
-    test_event = open(test_data + 'cis29.json')
-    event = json.loads(test_event.read())
-    test_event.close()
+    event = utils.load_test_data(test_data + 'cis29.json', my_region)
 
     # Mock Notifier
     init = mocker.patch('lib.sechub_findings.Finding.flag')
@@ -56,9 +58,8 @@ def test_event_good(mocker):
     #--------------------------
     # Test data
     #
-    test_event = open(test_data + 'cis29.json')
-    event = json.loads(test_event.read())
-    test_event.close()
+    event = utils.load_test_data(test_data + 'cis29.json', my_region)
+
     sns_message = {
         'Note': '"Enable VPC flow logging in all VPCs" remediation was successful',
         'State': 'RESOLVED',
@@ -134,15 +135,13 @@ def test_event_good(mocker):
     resolve.assert_called_once_with(
         'RESOLVED: "Enable VPC flow logging in all VPCs" remediation was successful'
     )
-    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, 'us-east-1')
+    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, my_region)
 
 def test_not_remediated(mocker):
     #--------------------------
     # Test data
     #
-    test_event = open(test_data + 'cis29.json')
-    event = json.loads(test_event.read())
-    test_event.close()
+    event = utils.load_test_data(test_data + 'cis29.json', my_region)
 
     os.environ['AWS_SESSION_TOKEN'] = 'FAKETOKEN'
     os.environ['FLOW_LOG_ROLE_ARN'] = 'FAKELOGROLEARN'