+++ source/playbooks/CIS/lambda/tests/test_cis43.py
@@ -27,10 +27,14 @@ from lib.logger import Logger
 from lib.applogger import LogHandler
 from lib.awsapi_helpers import BotoSession, AWSClient
 import lib.sechub_findings
+import tests.file_utilities as utils
 
 log_level = 'info'
 logger = Logger(loglevel=log_level)
 test_data = 'tests/test_data/'
+
+my_session = boto3.session.Session()
+my_region = my_session.region_name
 
 AWS = AWSClient()
 
@@ -38,9 +42,8 @@ def test_event_good(mocker):
     #--------------------------
     # Test data
     #
-    test_event = open(test_data + 'cis43.json')
-    event = json.loads(test_event.read())
-    test_event.close()
+    event = utils.load_test_data(test_data + 'cis43.json', my_region)
+
     sns_message = {
         'Note': '"Remove all rules from the default security group" remediation was successful',
         'State': 'RESOLVED',
@@ -160,4 +163,4 @@ def test_event_good(mocker):
     resolve.assert_called_once_with(
         'RESOLVED: "Remove all rules from the default security group" remediation was successful'
     )
-    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, 'us-east-1')
+    sns.assert_called_with('SO0111-SHARR_Topic', sns_message, my_region)